Separate Measurement and MeasurementContainer files
Theoretically MeasurementContainer can hold any class with a few properties;
the provided Measurement template is just one possibility. They are separate.
