+++ wave_containers/tests/containers/measurement_test.cpp
@@ -1,6 +1,8 @@
 #include "wave/wave_test.hpp"
 
+#include "wave/containers/measurement_container.hpp"
 #include "wave/containers/measurement.hpp"
+
 namespace wave {
 
 enum class SomeSensors { S1, S2, S3 };
